SPARK-3: The spark-shell fails under a dist.
Problem:
The spark-shell will fail to start when Spark is deployed using the
tar.gz file built by ./make-distribution. The problem is generated
by the spark-class shell that fails to setup a CLASSPATH if
Spark is a RELEASE.

Fix:
The script `spark-class` now sets the CLASSPATH if it is a RELEASE/dist
by pointing to the only assembly that should exist in the `./jars`

Environment:
*NIX environments. A similar change should be applied to
`spark-class.cmd`.

[ticket: SPARK-3] : https://issues.apache.org/jira/browse/SPARK-3